requirements.txt
flask==2.3.2
looker-sdk==23.10.0
python-dotenv
flask-cors=4.0.0